backend/requirements.txt
requests
pandas
numpy
scikit-learn
beautifulsoup4
lxml
Flask
Flask-Cors
python-dotenv
